% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_tab_dbase_fct_helpers.R
\name{append_data}
\alias{append_data}
\title{Appends rows into a table}
\usage{
append_data(conn, table, value)
}
\arguments{
\item{conn}{The connection object}

\item{table}{String, The name of the DB table}

\item{value}{The dataframe of values.  See DBI::dbAppendTable()}
}
\description{
Appends rows into a table
}
