% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_data_validation.R
\name{check_filetype}
\alias{check_filetype}
\title{Check that the file has the correct type}
\usage{
check_filetype(path)
}
\arguments{
\item{path}{The path of the file}
}
\description{
Check that the file has the correct type
}
