% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_dbase_operations.R
\name{check_statement_in_db}
\alias{check_statement_in_db}
\title{Check if a statemnt is already in the db based on the date}
\usage{
check_statement_in_db(statement_date, statement_type)
}
\description{
Check if a statemnt is already in the db based on the date
}
