% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_data_validation.R
\name{check_var_names}
\alias{check_var_names}
\title{Check that the data supplied has the correct variable names}
\usage{
check_var_names(dta, type)
}
\arguments{
\item{dta}{Tibble. The uploaded dataset?}

\item{type}{String length 1. One of `csi` or `ticket`}
}
\description{
Check that the data supplied has the correct variable names
}
