% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_data_validation.R
\name{check_var_names_bank}
\alias{check_var_names_bank}
\title{Check the column names of a loaded Bank dataset}
\usage{
check_var_names_bank(dta, bank)
}
\arguments{
\item{dta}{The loaded data}

\item{bank}{String length 1. The name of the bank.Takes the global value \code{banks}}
}
\value{
NULL if all ok, else a character vector of names that are missing
}
\description{
Check the column names of a loaded Bank dataset
}
