% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_data_validation.R
\name{check_var_names_SAGE}
\alias{check_var_names_SAGE}
\title{Check the column names of a loaded SAGE dataset}
\usage{
check_var_names_SAGE(dta)
}
\arguments{
\item{dta}{The loaded data}
}
\value{
NULL if all ok, else a character vector of names that are missing
}
\description{
Check the column names of a loaded SAGE dataset
}
