% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_tab_dbase_fct_helpers.R
\name{entry_form}
\alias{entry_form}
\title{Entry form to add a new store & email
a modalDialog. Needs to be a function to run properly (lazy loading?)}
\usage{
entry_form(session, edit = FALSE, store = NULL)
}
\description{
Entry form to add a new store & email
a modalDialog. Needs to be a function to run properly (lazy loading?)
}
