% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_tab_dbase_fct_helpers.R
\name{labelMandatory}
\alias{labelMandatory}
\title{Function to label mamdatory fields}
\usage{
labelMandatory(label)
}
\description{
This function will be used later on to mark any fields in the entry form that are mandatory.
Taken from
}
