% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_tab_load_fct_helpers.R
\name{load_bank}
\alias{load_bank}
\title{Load BANK statement}
\usage{
load_bank(path, bank)
}
\arguments{
\item{path}{The path to the file}

\item{bank}{String.One of c("HB", "BOC")}
}
\description{
Load BANK statement
}
