% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_read_helpers.R
\name{read_ticket_hour}
\alias{read_ticket_hour}
\title{Read in a Ticket Hour CSI}
\usage{
read_ticket_hour(path)
}
\arguments{
\item{path}{A path to a .csv file}
}
\description{
Read in a Ticket Hour CSI
}
