% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{safe_readXL}
\alias{safe_readXL}
\title{Safe read an excel file}
\usage{
safe_readXL(...)
}
\arguments{
\item{...}{temporary}
}
\description{
to construct it I need a "dummy"" output first and the
construct it in the .onLoad
see https://purrr.tidyverse.org/reference/faq-adverbs-export.html
We do it this way so that the safe_readXL is created on package load
to  get the latest update of purrr
}
